gazebo_gimbal_controller: work on vehicle yaw
We now get the vehicle yaw from the mavlink message. This means we are
independent of the PX4 orientation and can also test the gimbal device
on its own using the gimbal_device_tester from the MAVSDK examples.

A nicer solution in the future might be to have a very simple state
estimator implemented in the controller which then uses the
autopilot_state_to_gimbal_device message for corrections.

This commit also implements yaw follow and lock mode.